all_streamlit/pyproject.toml
[project]
name = "text2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0,<0.28.0",
    "openai>=1.60.0",
    "opensearch-py>=2.8.0",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.0",
    "sentence-transformers>=3.3.1",
    "sqlalchemy>=2.0.36",
    "streamlit>=1.41.1",
    "vanna[chromadb,ollama,postgres]>=0.7.5",
]
